Add namespace for styles.
